Clean up types and leftover todos
