+++ packages/server/lib/controllers/xhrs.ts
@@ -1,28 +1,11 @@
+import { parseContentType } from '@packages/net-stubbing/lib/server/util'
 import _ from 'lodash'
-import mime from 'mime'
 import Promise from 'bluebird'
 import fixture from '../fixture'
 
 const fixturesRe = /^(fx:|fixture:)/
-const htmlLikeRe = /<.+>[\s\S]+<\/.+>/
 
-const isValidJSON = function (text) {
-  if (_.isObject(text)) {
-    return true
-  }
-
-  try {
-    const o = JSON.parse(text)
-
-    return _.isObject(o)
-  } catch (error) {
-    false
-  }
-
-  return false
-}
-
-module.exports = {
+export = {
   handle (req, res, config, next) {
     const get = function (val, def?) {
       return decodeURI(req.get(val) || def)
@@ -37,7 +20,7 @@ module.exports = {
       // figure out the stream interface and pipe these
       // chunks to the response
       return this.getResponse(response, config)
-      .then((resp: { data: any, encoding?: string }) => {
+      .then((resp: { data: any, encoding?: BufferEncoding }) => {
         let { data, encoding } = resp
 
         // grab content-type from x-cypress-headers if present
@@ -115,22 +98,6 @@ module.exports = {
     return Promise.resolve({ data: resp })
   },
 
-  parseContentType (response) {
-    const ret = (type) => {
-      return mime.getType(type) //+ "; charset=utf-8"
-    }
-
-    if (isValidJSON(response)) {
-      return ret('json')
-    }
-
-    if (htmlLikeRe.test(response)) {
-      return ret('html')
-    }
-
-    return ret('text')
-  },
-
   parseHeaders (headers, response) {
     try {
       headers = JSON.parse(headers)
@@ -141,7 +108,7 @@ module.exports = {
     }
 
     if (headers['content-type'] == null) {
-      headers['content-type'] = this.parseContentType(response)
+      headers['content-type'] = parseContentType(response)
     }
 
     return headers
